requirements.txt
fastapi
uvicorn
jinja2
python-multipart
sqlmodel
psycopg2-binary
python-dotenv